chore: v0.1.1 — pack GribIndexEntry, harden FetchContent CI
GribIndexEntry reordered largest-align-first + record->uint32_t: zero
internal padding, sizeof 160->152, tighter vector scan. CI: authenticated
git rewrite so the five build-time github clones (glaze, googletest,
eccodes, libaec, ecbuild) aren't anonymous-rate-limit 403'd. Field-name
API unchanged; 2/2 tests, lint clean.

+++ src/models/sidecar.cpp
@@ -99,7 +99,7 @@ Result<std::vector<GribIndexEntry>> parse_nomads_idx(std::string_view text) {
 		e.raw = std::string(line);
 		std::uint64_t rec = 0;
 		if (to_u64(f[0], rec)) {
-			e.record = static_cast<std::size_t>(rec);
+			e.record = static_cast<std::uint32_t>(rec);
 		}
 		if (!to_u64(f[1], e.offset)) {
 			return std::unexpected(
